+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - **ENH: Add a function `snap.get_snap_version` to retrieve current SNAP version**  ([#172](https://github.com/sertit/eoreader/issues/172))
 - **ENH: Allow to pass directly a version object to `misc.compare_version`**
+- FIX: Prune empty keywords from kwargs in `rasters(_rio).write` to avoid throwing GDAL warnings/errors
 
 ## 1.41.0 (2024-08-06)